feat: Learnt props and propTypes
