--- Front-End/React JS/Code_With_Harry_Tutorial/Tutorial_1/textutils/src/App.js
@@ -1,31 +1,0 @@
-import './App.css';
-
-function App() {
-  return (
-    <>
-      <nav className="navbar navbar-expand-lg navbar-light bg-light">
-        <div className="container-fluid">
-          <a className="navbar-brand" href="/">Text Utils</a>
-          <button className="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-            <span className="navbar-toggler-icon"></span>
-          </button>
-          <div className="collapse navbar-collapse" id="navbarNav">
-            <ul className="navbar-nav">
-              <li className="nav-item">
-                <a className="nav-link active" aria-current="page" href="/">Home</a>
-              </li>
-              <li className="nav-item">
-                <a className="nav-link" href="/">About</a>
-              </li>
-              <li className="nav-item">
-                <a className="nav-link" href="/">Contact</a>
-              </li>
-            </ul>
-          </div>
-        </div>
-      </nav>
-    </>
-  );
-}
-
-export default App;
