+++ Front-End/React JS/Code_With_Harry_Tutorial/Tutorial_1/textutils/src/components/App/App.js
@@ -1,0 +1,12 @@
+import Navbar from '../Navbar/Navbar';
+
+function App() {
+  return (
+    <>
+      {/*<Navbar title="Text Utils"/>*/}
+      <Navbar />
+    </>
+  );
+}
+
+export default App;
