+++ Front-End/React JS/Code_With_Harry_Tutorial/Tutorial_1/textutils/src/index.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import App from './components/App/App';
 
 ReactDOM.render(
   <React.StrictMode>
@@ -10,4 +9,3 @@ ReactDOM.render(
   document.getElementById('root')
 );
 
-reportWebVitals();
